+++ apps/frontend/src/libs/components/components.ts
@@ -1,10 +1,13 @@
 export { AppHeader } from "./app-header/app-header.js";
 export { App } from "./app/app.js";
 export { Button } from "./button/button.js";
+export { DashboardWrapper } from "./dashboard-wrapper/dashboard-wrapper.js";
+export { Dashboard } from "./dashboard/dashboard.js";
 export { DecorativeImage } from "./decorative-image/decorative-image.js";
 export { Input } from "./input/input.js";
 export { Link } from "./link/link.js";
 export { Loader } from "./loader/loader.js";
+export { Logo } from "./logo/logo.js";
 export { ProtectedRoute } from "./protected-route/protected-route.js";
 export { RouterProvider } from "./router-provider/router-provider.js";
 export { Provider as StoreProvider } from "react-redux";