+++ apps/frontend/src/routes.tsx
@@ -1,12 +1,9 @@
 import { type RouteObject } from "react-router-dom";
 
+import { Dashboard, DashboardWrapper } from "./libs/components/components.js";
 import { AppRoute, RouteAccess } from "./libs/enums/enums.js";
-import {
-	Dashboard,
-	Plan,
-	Wrapper,
-} from "./pages/dashboard-wrapper-mock/components/components.js";
-import { Auth, Home, NotFound } from "./pages/pages.js";
+import { Plan } from "./pages/dashboard-wrapper-mock/components/components.js";
+import { Auth, ChooseStyle, Home, NotFound } from "./pages/pages.js";
 
 type CustomRouteObject = RouteObject & { handle: RouteHandle };
 
@@ -43,9 +40,14 @@ const routes: CustomRouteObject[] = [
 				path: AppRoute.PLAN,
 			},
 		],
-		element: <Wrapper />,
+		element: <DashboardWrapper />,
 		handle: { access: RouteAccess.AUTHENTICATED },
 		path: AppRoute.ROOT,
+	},
+	{
+		element: <ChooseStyle />,
+		handle: { access: RouteAccess.AUTHENTICATED },
+		path: AppRoute.CHOOSE_STYLE,
 	},
 	{
 		element: <NotFound />,